add dao (good2)

+++ src/main/java/com/phonebook/dao/EmployeeDao.java
@@ -9,9 +9,9 @@ public interface EmployeeDao {
 
     public void updateEmployee(Employee employee);
 
-    public void removeEmployee(int id);
+    public void removeEmployee(long id);
 
-    public Employee getEmployeeById(int id);
+    public Employee getEmployeeById(long id);
 
     public List<Employee> listEmployee();
